Cluster utitility functions

These functions provide useful helpers for performaning common operations.

cluster_assign(cluster, name, expr)

cluster_assign_each(cluster, name, values)

cluster_copy(cluster, names, env = caller_env())

cluster_get(cluster, name)

cluster_rm(cluster, names)

cluster_ls(cluster)

cluster_library(cluster, packages)

Arguments

cluster
Cluster to work on

name, names
Name of variable or variables, as strings.

expr
An expression to evaluate.

values
A list of values, one for each worker.

env
Environment in which to look for varibles to copy.

packages
Character vector of packages to load

Value

All functions that modify the worker environment invisibly return the cluster so they can chain them together. The other functions return lists with with one element for each worker.

Examples

cl <- new_cluster(2)
cl %>%
  cluster_assign("x", runif(1)) %>%
  cluster_get("x")

#> [[1]]
#> [1] 0.9167678
#> 
#> [[2]]
#> [1] 0.2200362
#> 

# Inline the expression using !! if you want to compute it locally
cl %>%
  cluster_assign("y", !!runif(1)) %>%
  cluster_get("x")

#> [[1]]
#> [1] 0.9167678
#> 
#> [[2]]
#> [1] 0.2200362
#> 

cl %>%
  cluster_assign_each("z", list(3, 4)) %>%
  cluster_get("z")

#> [[1]]
#> [1] 3
#> 
#> [[2]]
#> [1] 4
#> 

cl %>% cluster_ls()

#> [[1]]
#> [1] "data" "env"  "x"    "y"    "z"   
#> 
#> [[2]]
#> [1] "data" "env"  "x"    "y"    "z"   
#> 

w <- 10
cluster_copy(cl, "w")

cl %>% cluster_rm(c("w", "x", "y", "z"))
cl %>% cluster_ls()

#> [[1]]
#> [1] "data" "env" 
#> 
#> [[2]]
#> [1] "data" "env" 
#> 

cl %>% cluster_call(search())

#> [[1]]
#>  [1] ".GlobalEnv"        "package:stats"     "package:graphics" 
#>  [4] "package:grDevices" "package:utils"     "package:datasets" 
#>  [7] "package:methods"   "Autoloads"         "tools:callr"      
#> [10] "package:base"     
#> 
#> [[2]]
#>  [1] ".GlobalEnv"        "package:stats"     "package:graphics" 
#>  [4] "package:grDevices" "package:utils"     "package:datasets" 
#>  [7] "package:methods"   "Autoloads"         "tools:callr"      
#> [10] "package:base"     
#> 

cl %>%
  cluster_library("magrittr") %>%
  cluster_call(search())

#> 
#> Attaching package: ‘magrittr’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     equals, is_less_than, not

#> [[1]]
#>  [1] ".GlobalEnv"        "package:magrittr"  "package:stats"    
#>  [4] "package:graphics"  "package:grDevices" "package:utils"    
#>  [7] "package:datasets"  "package:methods"   "Autoloads"        
#> [10] "tools:callr"       "package:base"     
#> 
#> [[2]]
#>  [1] ".GlobalEnv"        "package:magrittr"  "package:stats"    
#>  [4] "package:graphics"  "package:grDevices" "package:utils"    
#>  [7] "package:datasets"  "package:methods"   "Autoloads"        
#> [10] "tools:callr"       "package:base"     
#>